Extract GitHubIssueFetcher to its own file
